+++ hello_world/plugin_001.ml
@@ -1,4 +1,4 @@
-open Core.Std let _ = _squelch_unused_module_warning_
+open! Core.Std
 
 (* example of plugin of type Plugin_intf.S *)