+++ hello_world/plugin_002.ml
@@ -1,3 +1,3 @@
-open Core.Std let _ = _squelch_unused_module_warning_
+open! Core.Std
 
 let message = 42