+++ hello_world/plugin_003.ml
@@ -1,4 +1,4 @@
-open Core.Std let _ = _squelch_unused_module_warning_
+open! Core.Std
 
 let f s = s^s^s