+++ myocamlbuild.ml
@@ -2,6 +2,27 @@
 (* OASIS_STOP *)
 # 4 "myocamlbuild.ml"
 
+(* Temporary hacks *)
+let js_hacks = function
+  | After_rules ->
+    rule "Generate a cmxs from a cmxa"
+      ~dep:"%.cmxa"
+      ~prod:"%.cmxs"
+      ~insert:`top
+      (fun env _ ->
+         Cmd (S [ !Options.ocamlopt
+                ; A "-shared"
+                ; A "-linkall"
+                ; A "-I"; A (Pathname.dirname (env "%"))
+                ; A (env "%.cmxa")
+                ; A "-o"
+                ; A (env "%.cmxs")
+            ]));
+
+    (* Pass -predicates to ocamldep *)
+    pflag ["ocaml"; "ocamldep"] "predicate" (fun s -> S [A "-predicates"; A s])
+  | _ -> ()
+
 let dispatch = function
   | Before_options ->
     Options.make_links := false
@@ -35,4 +56,9 @@ let dispatch = function
   | _ ->
     ()
 
-let () = Ocamlbuild_plugin.dispatch (fun hook -> dispatch hook; dispatch_default hook)
+let () =
+  Ocamlbuild_plugin.dispatch (fun hook ->
+    js_hacks hook;
+    Ppx_driver_ocamlbuild.dispatch hook;
+    dispatch hook;
+    dispatch_default hook)