+++ src/params.ml
@@ -1,5 +1,8 @@
-let build_info_available = false
+(* In the public release, this module is replaced by:
 
-let build_info_as_sexp = Sexplib.Sexp.Atom "no_build_info"
-let version = "(no version info)"
-let build_info = "(no build info)"
+let build_info_as_sexp =
+  Sexplib.Sexp.Atom (Digest.to_hex (Digest.file Sys.executable_name))
+let version = "NO_VERSION_UTIL"
+
+*)
+include Core.Std.Version_util