+++ src/shell.ml
@@ -49,7 +49,7 @@ let endline std = if std = "" then std else std ^ "\n"
 
 let make_run from_output ?working_dir ?(quiet_or_error = false) prog args =
   let command_text =
-    lazy (prog::args |> <:sexp_of< string list >> |> Sexp.to_string)
+    lazy (prog::args |> [%sexp_of: string list] |> Sexp.to_string)
   in
   if !echo then
     Core.Std.Printf.printf "Shell: %s\n%!" (force command_text);
@@ -84,7 +84,7 @@ let make_run from_output ?working_dir ?(quiet_or_error = false) prog args =
         let status =
           match status with
           | Some status -> Sexp.to_string
-            (<:sexp_of< Core.Std.Unix.Exit_or_signal.error >> status)
+            ([%sexp_of: Core.Std.Unix.Exit_or_signal.error] status)
           | None -> "error trace on stdout or stderr"
         in
         sprintf "working_dir: %s\nstatus: %s\ncommand: %s\n%s%s"