+++ test/test_lib/test.ml
@@ -1,2 +1,2 @@
-TEST = true
+let%test _ = true
 (* just testing that the linking works *)