+++ test/test_with_sexp_dep_ppx_style.mli
@@ -1,0 +1,7 @@
+open Core.Std
+
+type t [@@deriving sexp]
+
+val sexped : Sexp.t
+
+val first : t