+++ test/test_with_sexp_ppx_style.mli
@@ -1,0 +1,8 @@
+open Core.Std
+
+type t [@@deriving sexp]
+
+val sexped : Sexp.t
+
+val first : t
+